Deals cards, broken gameplay
Okay, so I messed around with the game play. The war() method needs to fixed as does the turn() method. It deals correctly but these methods are used in the comparisons. I know that it is a lot of change, so let me know if you are going to completely redo anything haha. It is about 11 now so I am going to sleep, I'll work on it more tomorrow!